refactor serde guard yaml module
